# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation ",
        "description": "API Documentation",
        "contact": {
            "email": "srundavith@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get List movies Data",
                "description": "enter your movies here",
                "operationId": "movies",
                "responses": {
                    "default": {
                        "description": "return array model movies"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Create a new movie",
                "description": "Create a new movie with various attributes including file uploads for poster and cover images.",
                "operationId": "createMovie",
                "requestBody": {
                    "description": "Form data for creating a movie",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "overview",
                                    "run_time",
                                    "release_date",
                                    "total_likes",
                                    "total_ratings",
                                    "average_rating",
                                    "poster_image",
                                    "cover_image",
                                    "trailer_url",
                                    "last_upload_date",
                                    "subscription_only",
                                    "expired_subscription_only"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "overview": {
                                        "type": "string"
                                    },
                                    "run_time": {
                                        "type": "integer"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_likes": {
                                        "type": "integer"
                                    },
                                    "total_ratings": {
                                        "type": "integer"
                                    },
                                    "average_rating": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "poster_image_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "poster_image_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "cover_image_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "cover_image_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "trailer_url": {
                                        "type": "string"
                                    },
                                    "last_upload_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "subscription_only": {
                                        "description": "Boolean represented as 1 (true) or 0 (false)",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "expired_subscription_only": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "Movies/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update movies",
                "description": "-",
                "operationId": "movies/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "overview",
                                    "run_time",
                                    "release_date",
                                    "total_likes",
                                    "total_ratings",
                                    "average_rating",
                                    "poster_image",
                                    "cover_image",
                                    "trailer_url",
                                    "last_upload_date",
                                    "subscription_only",
                                    "expired_subscription_only"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "overview": {
                                        "type": "string"
                                    },
                                    "run_time": {
                                        "type": "integer"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_likes": {
                                        "type": "integer"
                                    },
                                    "total_ratings": {
                                        "type": "integer"
                                    },
                                    "average_rating": {
                                        "type": "string"
                                    },
                                    "poster_image": {
                                        "type": "string"
                                    },
                                    "cover_image": {
                                        "type": "string"
                                    },
                                    "trailer_url": {
                                        "type": "string"
                                    },
                                    "last_upload_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "subscription_only": {
                                        "type": "string"
                                    },
                                    "expired_subscription_only": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete movies",
                "description": "-",
                "operationId": "movies/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/movie_videos": {
            "post": {
                "tags": [
                    "Movie_Videos"
                ],
                "summary": "Create a new movie video",
                "description": "Create a new movie video with either a file upload or a URL.",
                "operationId": "createMovieVideo",
                "requestBody": {
                    "description": "Movie video data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "movie_id",
                                    "type",
                                    "official",
                                    "subscription"
                                ],
                                "properties": {
                                    "movie_id": {
                                        "type": "string"
                                    },
                                    "video_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "video_url": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "season_number": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "episode_number": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "part_number": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "movie",
                                            "trailer"
                                        ]
                                    },
                                    "official": {
                                        "description": "Boolean represented as 1 (true) or 0 (false)",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "subscription": {
                                        "description": "Boolean represented as 1 (true) or 0 (false)",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "subscription_start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "subscription_end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model movie-videos"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/movie-videos/{id}": {
            "get": {
                "tags": [
                    "Movie_Videos"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "movie-videos/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movie_Videos"
                ],
                "summary": "Update movie-videos",
                "description": "-",
                "operationId": "movie-videos/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "video"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "video": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie_Videos"
                ],
                "summary": "Delete movie-videos",
                "description": "-",
                "operationId": "movie-videos/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}